Remove some of the implicit cast in upnp part
Remove some of the "implicit integer or enum conversions" as well as
some access to NULL reference in upnp part.

+++ upnp/inc/upnp.h
@@ -61,8 +61,8 @@
 	/* Other systems ??? */
 #endif
 
-#define LINE_SIZE  180
-#define NAME_SIZE  256
+#define LINE_SIZE  (size_t)180
+#define NAME_SIZE  (size_t)256
 #define MNFT_NAME_SIZE  64
 #define MODL_NAME_SIZE  32
 #define SERL_NUMR_SIZE  64
